LIB-446 fix(dialog): rename simpledialog to confirmdialog
