+++ apps/storybook/src/App.vue
@@ -1,13 +1,13 @@
 <script setup lang="ts">
 import { RouterLink, RouterView } from 'vue-router'
 import HelloWorld from './components/HelloWorld.vue'
-import { FzSimpleDialog } from '@fiscozen/dialog'
+import { FzConfirmDialog } from '@fiscozen/dialog'
 </script>
 
 <template>
   <header>
     <img alt="Vue logo" class="logo" src="@/assets/logo.svg" width="125" height="125" />
-    <FzSimpleDialog title="Il mio titolo" confirm-label="ok" cancel-label="ko"></FzSimpleDialog>
+    <FzConfirmDialog title="Il mio titolo" confirm-label="ok" cancel-label="ko"></FzConfirmDialog>
 
     <div class="wrapper">
       <HelloWorld msg="You did it!" />
