+++ apps/storybook/src/stories/Dialog.stories.ts
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/vue3'
 import { ref } from 'vue'
 
-import { FzDialog, FzSimpleDialog, FzDialogProps } from '@fiscozen/dialog'
+import { FzDialog, FzConfirmDialog, FzDialogProps } from '@fiscozen/dialog'
 import {FzButton} from '@fiscozen/button'
 
 // More on how to set up stories at: https://storybook.js.org/docs/writing-stories
@@ -22,14 +22,14 @@ type Story = StoryObj<typeof meta>
 
 const simpleDialog = (args : FzDialogProps) => ({
   setup() {
-    const dialog = ref<InstanceType<typeof FzSimpleDialog>>();
+    const dialog = ref<InstanceType<typeof FzConfirmDialog>>();
     return { args, dialog: dialog.value}
   },
-  components: { FzDialog, FzSimpleDialog, FzButton },
+  components: { FzDialog, FzConfirmDialog, FzButton },
   template: `
     <div class="w-screen h-screen">
       <FzButton @click="dialog.show()">Open Modal</FzButton>
-      <FzSimpleDialog v-bind="args" ref="dialog"></FzSimpleDialog>
+      <FzConfirmDialog v-bind="args" ref="dialog"></FzConfirmDialog>
     </div>
   `
 })
