+++ packages/dialog/src/index.ts
@@ -1,3 +1,3 @@
 export { default as FzDialog } from "./FzDialog.vue";
-export { default as FzSimpleDialog } from "./FzSimpleDialog.vue";
+export { default as FzConfirmDialog } from "./FzConfirmDialog.vue";
 export * from "./types";
